- BUG: Filemaneger no extended search - BUG: rights dropdown
git-svn-id: file:///Users/jan/Dropbox/svn/FlexyAdmin/trunk@3716 0c02b2c5-a8bd-4e35-85b2-cf8cacb96951

Former-commit-id: 41083005e1ad233ab7bd8a7b261c55b207048f97
Former-commit-id: 5aeef56b10b36e9b0ea1cce0dc50a57c8e0cafe6
Former-commit-id: f8f99d5d21c9ab2d3ba44c16ba34b371e938304a [formerly 00cc9b9171c059e5be26993a33329a354c7c05a5]
Former-commit-id: 7b18c0c005fe242c8a317174fd5ca2f1063b188a

+++ sys/flexyadmin/models/data/Res_media_files.php
@@ -44,10 +44,12 @@ Class Res_media_files extends Data_Core {
    * @author Jan den Besten
    */
   public function _delete_files( $deleted_data ) {
-    foreach ($deleted_data as $id => $file_row) {
-      $this->file_manager->set_path( $file_row['path'] );
-      if ( !$this->file_manager->delete_file( $file_row['file'] ) ) {
-        unset($deleted_data[$id]);
+    if (is_array($deleted_data)) {
+      foreach ($deleted_data as $id => $file_row) {
+        $this->file_manager->set_path( $file_row['path'] );
+        if ( !$this->file_manager->delete_file( $file_row['file'] ) ) {
+          unset($deleted_data[$id]);
+        }
       }
     }
     return $deleted_data;
